finished ttf build

+++ scripts/ttfbuild.py
@@ -22,10 +22,10 @@ INFO='''<?xml version="1.0" encoding="UTF-8"?>
 for file in file_list:
     path='instances/'+file+'/font.ufo'
     print "[INFO] Checking out lines and resetting UFO"
-    subprocess.call(['checkOutlinesUFO', '-wd', '-e', path])
-
+    subprocess.call(['checkOutlinesUFO', '-e', '-all','-wd', path])
+	
     for item in os.listdir(path):
-        if os.path.isfile(path+'/'+item) and not item=="lib.plist":
+        if os.path.isfile(path+'/'+item) and not item=="lib.plist" and not item=="metainfo.plist":
             os.remove(path+'/'+item)
 
     with open(path+'/features.fea', 'w') as f:
